Add files via upload
Tasty Treats is a modern restaurant that specializes in delivering delicious meals across Bangalore. With a wide variety of dishes and reliable service. Tasty Treats brings the joy of great food right to your doorstep. While it is a dine-in restaurant, its main focus is on serving online food orders, ensuring customers can enjoy fresh and tasty meals at home. Tasty Treats has identified order cancellations as a key area for improvement in its online food delivery service. While cancellations due to changes in customers' plans are beyond the company's control, it is committed to addressing other contributing factors. For instance, if an order is cancelled due to an item being out of stock, the system automatically detects and flags the issue. By resolving such problems, Tasty Treats aims to enhance the customer experience, build stronger trust, and foster greater loyalty. The company is providing the order details of its top customers for analysis to better understand and address these issues.