Add files via upload
We are planning a marketing campaign in the upcoming IPL season which will have a significant cost impact for us. We also forecast that it will boost the sales in the majority of our city as they have an IPL fanbase. For this purpose we have asked our regional managers to provide the forecast for both of our products “NutriKid” and “ZestVita” for both cases (IPL marketing is done and IPL marketing is not done).